Create removeDuplicates
works by checking the first value of the list and checking the list to see if the count of that value is 1. If it is you increment start up one. Once start == length of array, you check the length of the array to get the answer leetcode wants. Leetcode wants you to not create another array and needs use to remove as you go. My Solution is horrible, looked at solution and they are using two pointer, need to read up on that