Two pointers
Better approach is to do two pointers. One pointer is the start pointer or slow. It checks the first variable in the array and compares it to the second. If the variables are different we increment start by 1 and change nums[start] to current number. This changes the array in place. This is much faster than my last attempt.