two pointer solution
Lets say you can't change the length of the list. The only possible solution would than be two pointers. In this case start begins at 0 index and whenever nums[start] != val, start is incremented one. If not nums[x] keeps going but start starts the same. This allows you to move down the array and change the array values in order.