https://leetcode.com/problems/uncommon-words-from-two-sentences/
make the strings into a liist, than into a set. Allow a set allows you to use .difference, which tells you which is in one list and not the other. Than you make dictionaries to determine if that unique word is only in the word one and append it to ans.