Update sortArrayByParity
Using unpacking allows you to rearrange the array in place, instead of making more arrays. This is probably the best solution. It works by going from both ends. If % 2 of start == 1, it means the beginning is odd and you want to move this odd number to the back, likewise if nums[end] is even, you want to move it in front. The only way to do this is unpacking. You can't do it in two separate lines. The other if statements test if the number is even it goes to next number and if the end is odd it substracts a number.