Create 2078.py
make two different lists, one for possible numbers to check, the other for the first index it occurs. Iterable backwards and when possible numbers isn't the last number, you will check if that the max. Sounds like this solution would be O(n^4). Solution seems kind of convoluted, will try to reformat....