Medium
make a dictionary, find the max freq, and append the max freq to a list, than subtract one from max freq. Return the answer as a string.
Since I used one while loop and one for loop, it appears to be run in O(N^4). Will try to think of a way to do it in one loop....