Easy
This appears to be a stack problem, you append the thing your questioning onto a stack and compare it and than pop it off and continue. Basically if stack is empty push a value in. If the next value is greater than the previous value, you would pop off the stack and append that value back to the stack. If the next value is less than the item in the stack, add one to ans.