SLIDING WINDOWS IS KIND OF HARD
my previous solution, did it using a stack, but apparently that is pretty slow and there is a better solution. It's called a sliding window. Basically the concept is you make to pointers left and right. You move the right until a duplicate char appears. When the duplicate appears, you stop and start removing from the front until you have removed the s[r] element. This allows you to appear the s[r] unit and still follow the rules of no duplicate units. THIS IS MUCH better than brute force, which I was using before. As it allows you to count units you already know are not a duplicate. It only took me 5 hours to understand this LOOOOOL